@@ -1,14 +1,14 @@
 import React from "react";
 
 function Header() {
-    return (
-        <section id="header">     
-            <div>       
-            <h1>I'm Subhadeep</h1>
-            <h2>a programmer</h2>
-            </div>            
-        </section>
-    );
+  return (
+    <section id="header">
+      <div>
+        <h1>I'm Subhadeep</h1>
+        <h2>a programmer</h2>
+      </div>
+    </section>
+  );
 }
 
 export default Header;
